working algorithms

+++ src/pillar.cpp
@@ -32,7 +32,7 @@ void Pillar::change_state(color_state state_change) {
 }
 
 void Pillar::draw(int x, int y, int width, int height) {
-  Rectangle r = {x, y, width, height};
+  Rectangle r = {(float) x, (float) y, (float) width, (float) height};
   
   DrawRectangleLines(x, y, width, height, BORDER_COLOR);
   DrawRectangle(x, y, width, height, _color);
